Moves Print memory from the cli to the display
The debug screen window enlarges from 800x600 to 1024x600. The complete
memory (4096 bytes) displays in a 256 lines of 16 one-byte elements fashion that
can be scrolled with the up/down arrows one line at a time.
The labels indicating from which section the bytes are, are removed until
the near future.