Modifies the text rendering
This is SOOOO slow it needs to be fixed to be merged. The change is that,
unlike the previous text rendering that copied the 10x10 char one at a
time, every cycle, this one prints everything on another surface/texture
(the size of the screen) that is copied to the screen one time every
cycle. In theory this is way faster than the former when there is a lot of
text to render but something is not working properly.